requirements.txt
# Autogenerate by build_requirements
utils_base-nuuuwan
utils_www-nuuuwan
openai
mistral